BLADERUNNER: Remove use of C99 math
C++ math functions are overloaded so operate using single-precision
when receiving a float input. The C standard library on FreeMiNT
does not fully support C99 math so use of sqrtf, sinf, etc.
instead of the C++ API does not work.

+++ engines/bladerunner/view.cpp
@@ -51,7 +51,7 @@ void View::setFovX(float fovX) {
 	_viewportHalfWidth = 320.0f;
 	_viewportHalfHeight = 240.0f;
 
-	_viewportDistance = 320.0f / tanf(_fovX / 2.0f);
+	_viewportDistance = 320.0f / tan(_fovX / 2.0f);
 }
 
 void View::calculateSliceViewMatrix() {
